Remove a level of indirection
